Reorganize PageMethod tests
