+++ tests/test_page_methods.py
@@ -1,5 +1,7 @@
 import logging
 import platform
+import subprocess
+from tempfile import NamedTemporaryFile
 
 import pytest
 from scrapy import Spider, Request
@@ -9,6 +11,15 @@ from scrapy_playwright.page import PageMethod
 
 from tests import make_handler
 from tests.mockserver import StaticMockServer
+
+
+def get_mimetype(file):
+    return subprocess.run(
+        ["file", "--mime-type", "--brief", file.name],
+        stdout=subprocess.PIPE,
+        universal_newlines=True,
+        check=False,
+    ).stdout.strip()
 
 
 @pytest.mark.asyncio
@@ -81,6 +92,102 @@ class MixinPageMethodTestCase:
         ) in caplog.record_tuples
         assert not req.meta["playwright_page_methods"]["is_closed"].result
         assert req.meta["playwright_page_methods"]["title"].result == "Awesome site"
+
+    @pytest.mark.asyncio
+    async def test_page_method_navigation(self):
+        async with make_handler({"PLAYWRIGHT_BROWSER_TYPE": self.browser_type}) as handler:
+            with StaticMockServer() as server:
+                req = Request(
+                    url=server.urljoin("/index.html"),
+                    meta={
+                        "playwright": True,
+                        "playwright_page_methods": [PageMethod("click", "a.lorem_ipsum")],
+                    },
+                )
+                resp = await handler._download_request(req, Spider("foo"))
+
+            assert isinstance(resp, HtmlResponse)
+            assert resp.request is req
+            assert resp.url == server.urljoin("/lorem_ipsum.html")
+            assert resp.status == 200
+            assert "playwright" in resp.flags
+            assert resp.css("title::text").get() == "Lorem Ipsum"
+            text = resp.css("p::text").get()
+            assert text == "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
+
+    @pytest.mark.asyncio
+    async def test_page_method_infinite_scroll(self):
+        async with make_handler({"PLAYWRIGHT_BROWSER_TYPE": self.browser_type}) as handler:
+            with StaticMockServer() as server:
+                req = Request(
+                    url=server.urljoin("/scroll.html"),
+                    headers={"User-Agent": "scrapy-playwright"},
+                    meta={
+                        "playwright": True,
+                        "playwright_page_methods": [
+                            PageMethod("wait_for_selector", selector="div.quote"),
+                            PageMethod(
+                                "evaluate", "window.scrollBy(0, document.body.scrollHeight)"
+                            ),
+                            PageMethod("wait_for_selector", selector="div.quote:nth-child(11)"),
+                            PageMethod(
+                                "evaluate", "window.scrollBy(0, document.body.scrollHeight)"
+                            ),
+                            PageMethod("wait_for_selector", selector="div.quote:nth-child(21)"),
+                        ],
+                    },
+                )
+                resp = await handler._download_request(req, Spider("foo"))
+
+            assert isinstance(resp, HtmlResponse)
+            assert resp.request is req
+            assert resp.url == server.urljoin("/scroll.html")
+            assert resp.status == 200
+            assert "playwright" in resp.flags
+            assert len(resp.css("div.quote")) == 30
+
+    @pytest.mark.asyncio
+    async def test_page_method_screenshot(self):
+        async with make_handler({"PLAYWRIGHT_BROWSER_TYPE": self.browser_type}) as handler:
+            with NamedTemporaryFile(mode="w+b") as png_file:
+                with StaticMockServer() as server:
+                    req = Request(
+                        url=server.urljoin("/index.html"),
+                        meta={
+                            "playwright": True,
+                            "playwright_page_methods": {
+                                "png": PageMethod("screenshot", path=png_file.name, type="png"),
+                            },
+                        },
+                    )
+                    await handler._download_request(req, Spider("foo"))
+
+                png_file.file.seek(0)
+                assert png_file.file.read() == req.meta["playwright_page_methods"]["png"].result
+                assert get_mimetype(png_file) == "image/png"
+
+    @pytest.mark.asyncio
+    async def test_page_method_pdf(self):
+        if self.browser_type != "chromium":
+            pytest.skip("PDF generation is supported only in Chromium")
+
+        async with make_handler({"PLAYWRIGHT_BROWSER_TYPE": self.browser_type}) as handler:
+            with NamedTemporaryFile(mode="w+b") as pdf_file:
+                with StaticMockServer() as server:
+                    req = Request(
+                        url=server.urljoin("/index.html"),
+                        meta={
+                            "playwright": True,
+                            "playwright_page_methods": {
+                                "pdf": PageMethod("pdf", path=pdf_file.name),
+                            },
+                        },
+                    )
+                    await handler._download_request(req, Spider("foo"))
+
+                pdf_file.file.seek(0)
+                assert pdf_file.file.read() == req.meta["playwright_page_methods"]["pdf"].result
+                assert get_mimetype(pdf_file) == "application/pdf"
 
 
 class TestPageMethodChromium(MixinPageMethodTestCase):
