+++ tests/test_playwright_requests.py
@@ -1,10 +1,8 @@
 import json
 import logging
 import platform
-import subprocess
 import sys
 from ipaddress import ip_address
-from tempfile import NamedTemporaryFile
 from unittest.mock import MagicMock, patch
 
 import pytest
@@ -21,15 +19,6 @@ from scrapy_playwright.page import PageMethod
 
 from tests import make_handler
 from tests.mockserver import MockServer, StaticMockServer
-
-
-def get_mimetype(file):
-    return subprocess.run(
-        ["file", "--mime-type", "--brief", file.name],
-        stdout=subprocess.PIPE,
-        universal_newlines=True,
-        check=False,
-    ).stdout.strip()
 
 
 class DialogSpider(Spider):
@@ -76,59 +65,6 @@ class MixinTestCase:
             assert resp.status == 200
             assert "playwright" in resp.flags
             assert "Request body: foo=bar" in resp.text
-
-    @pytest.mark.asyncio
-    async def test_page_method_navigation(self):
-        async with make_handler({"PLAYWRIGHT_BROWSER_TYPE": self.browser_type}) as handler:
-            with StaticMockServer() as server:
-                req = Request(
-                    url=server.urljoin("/index.html"),
-                    meta={
-                        "playwright": True,
-                        "playwright_page_methods": [PageMethod("click", "a.lorem_ipsum")],
-                    },
-                )
-                resp = await handler._download_request(req, Spider("foo"))
-
-            assert isinstance(resp, HtmlResponse)
-            assert resp.request is req
-            assert resp.url == server.urljoin("/lorem_ipsum.html")
-            assert resp.status == 200
-            assert "playwright" in resp.flags
-            assert resp.css("title::text").get() == "Lorem Ipsum"
-            text = resp.css("p::text").get()
-            assert text == "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
-
-    @pytest.mark.asyncio
-    async def test_page_method_infinite_scroll(self):
-        async with make_handler({"PLAYWRIGHT_BROWSER_TYPE": self.browser_type}) as handler:
-            with StaticMockServer() as server:
-                req = Request(
-                    url=server.urljoin("/scroll.html"),
-                    headers={"User-Agent": "scrapy-playwright"},
-                    meta={
-                        "playwright": True,
-                        "playwright_page_methods": [
-                            PageMethod("wait_for_selector", selector="div.quote"),
-                            PageMethod(
-                                "evaluate", "window.scrollBy(0, document.body.scrollHeight)"
-                            ),
-                            PageMethod("wait_for_selector", selector="div.quote:nth-child(11)"),
-                            PageMethod(
-                                "evaluate", "window.scrollBy(0, document.body.scrollHeight)"
-                            ),
-                            PageMethod("wait_for_selector", selector="div.quote:nth-child(21)"),
-                        ],
-                    },
-                )
-                resp = await handler._download_request(req, Spider("foo"))
-
-            assert isinstance(resp, HtmlResponse)
-            assert resp.request is req
-            assert resp.url == server.urljoin("/scroll.html")
-            assert resp.status == 200
-            assert "playwright" in resp.flags
-            assert len(resp.css("div.quote")) == 30
 
     @pytest.mark.asyncio
     async def test_timeout_value(self):
@@ -210,49 +146,6 @@ class MixinTestCase:
                 await req_handler(route, playwright_request)
 
     @pytest.mark.asyncio
-    async def test_page_method_screenshot(self):
-        async with make_handler({"PLAYWRIGHT_BROWSER_TYPE": self.browser_type}) as handler:
-            with NamedTemporaryFile(mode="w+b") as png_file:
-                with StaticMockServer() as server:
-                    req = Request(
-                        url=server.urljoin("/index.html"),
-                        meta={
-                            "playwright": True,
-                            "playwright_page_methods": {
-                                "png": PageMethod("screenshot", path=png_file.name, type="png"),
-                            },
-                        },
-                    )
-                    await handler._download_request(req, Spider("foo"))
-
-                png_file.file.seek(0)
-                assert png_file.file.read() == req.meta["playwright_page_methods"]["png"].result
-                assert get_mimetype(png_file) == "image/png"
-
-    @pytest.mark.asyncio
-    async def test_page_method_pdf(self):
-        if self.browser_type != "chromium":
-            pytest.skip("PDF generation is supported only in Chromium")
-
-        async with make_handler({"PLAYWRIGHT_BROWSER_TYPE": self.browser_type}) as handler:
-            with NamedTemporaryFile(mode="w+b") as pdf_file:
-                with StaticMockServer() as server:
-                    req = Request(
-                        url=server.urljoin("/index.html"),
-                        meta={
-                            "playwright": True,
-                            "playwright_page_methods": {
-                                "pdf": PageMethod("pdf", path=pdf_file.name),
-                            },
-                        },
-                    )
-                    await handler._download_request(req, Spider("foo"))
-
-                pdf_file.file.seek(0)
-                assert pdf_file.file.read() == req.meta["playwright_page_methods"]["pdf"].result
-                assert get_mimetype(pdf_file) == "application/pdf"
-
-    @pytest.mark.asyncio
     async def test_event_handler_dialog_callable(self):
         async with make_handler({"PLAYWRIGHT_BROWSER_TYPE": self.browser_type}) as handler:
             with StaticMockServer() as server:
@@ -262,6 +155,7 @@ class MixinTestCase:
                     meta={
                         "playwright": True,
                         "playwright_page_methods": [
+                            # trigger an alert
                             PageMethod("evaluate", "alert('foobar');"),
                         ],
                         "playwright_page_event_handlers": {
@@ -283,6 +177,7 @@ class MixinTestCase:
                     meta={
                         "playwright": True,
                         "playwright_page_methods": [
+                            # trigger an alert
                             PageMethod("evaluate", "alert('foobar');"),
                         ],
                         "playwright_page_event_handlers": {
